requirements.txt
joblib==1.5.1
lap==0.5.12
loguru==0.7.3
numpy==1.24.4
opencv-python==4.11.0.86
paho-mqtt==2.1.0
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.13.1
threadpoolctl==3.6.0
tqdm==4.67.1
